Move readableDuration for report layer
